BugFix with arg objects
Was originally written for a different returnObject from
ModelWriter.readForm().  That function changed, so this one had to as
well.
'arg' has been replaced with 'name' for the variable containing the name
of the argument.
'for (arg in' has been replaced with 'for (arg of', grabbing the object
instead of the index in the array.
Also an extra 'var' has been added 'for (var arg of args)' because it
should have been there already.